sqlite update

+++ bot.py
@@ -2,6 +2,7 @@
     Запуск бота
 '''
 import datetime
+from datetime import datetime, time, timedelta
 import shutil
 import sys
 
@@ -14,13 +15,18 @@ from get_tgd import get_data_students
 from get_tgd_teachers import get_data_teachers
 from db import create_connection, execute_query, execute_read_query
 
-from dotenv import load_dotenv
+from dotenv import dotenv_values
 from pathlib import Path
 import os
 from logging import getLogger, StreamHandler, Formatter, INFO
 import pandas as pd
 
 import platform
+
+import asyncio
+import aioschedule
+import nest_asyncio
+nest_asyncio.apply()
 system = platform.system()
 
 
@@ -31,11 +37,9 @@ handler = StreamHandler(stream=sys.stdout)
 handler.setFormatter(Formatter(fmt='[%(asctime)s: %(levelname)s] %(message)s'))
 logger.addHandler(handler)
 
-load_dotenv()
-env_path = Path('.')/'.env'
-load_dotenv(dotenv_path=env_path)
-TOKEN = os.getenv("TOKEN")
-ADMIN_KEY = os.getenv("ADMIN_KEY")
+config = dict(dotenv_values(".env") )
+TOKEN = config["TOKEN"]
+ADMIN_KEY = config["ADMIN_KEY"]
 
 bot = Bot(token=TOKEN, parse_mode=types.ParseMode.HTML)
 
@@ -46,6 +50,10 @@ data = {}
 
 data_teachers = {}
 connection = None
+# ТОЛЬКО ДЛЯ ПОСЛЕДНЕГО ОБНОВЛЕНИЯ
+# ЖЕЛАТЕЛЬНО ПОМЕНЯТЬ НА execute_query и execute_read_query потом
+
+cursor_for_update = None
 
 user_params = [
     'id',
@@ -54,50 +62,58 @@ user_params = [
     'word',
     'is_teacher',
     'teacher_last_name',
-    'is_admin'
+    'is_admin',
+    'notification_time'
 ]
 days = ['Понедельник', 'Вторник', 'Среда', 'Четверг', 'Пятница']
+
+interval_start_hour, interval_start_minute = 6, 0 # 6:00
+interval_end_hour, interval_end_minute = 11, 0 # 11:00
+interval_start, interval_end = "6:00",  "11:00"
+
 time_of_lesson = [
     {
         "start": "8:30",
         "end": "9:15",
     },
     {
-        "start": "9:30",
-        "end": "10:15",
+        "start": "9:35",
+        "end": "10:20",
     },
     {
-        "start": "10:30",
-        "end": "11:15",
+        "start": "10:40",
+        "end": "11:25",
     },
     {
-        "start": "11:30",
-        "end": "12:15",
+        "start": "11:45",
+        "end": "12:30",
     },
     {
-        "start": "12:35",
-        "end": "13:20",
+        "start": "12:50",
+        "end": "13:35",
     },
     {
-        "start": "13:40",
-        "end": "14:25",
+        "start": "13:55",
+        "end": "14:40",
     },
     {
-        "start": "14:45",
-        "end": "15:30",
+        "start": "15:00",
+        "end": "15:45",
     },
     {
-        "start": "15:40",
-        "end": "16:25",
+        "start": "15:55",
+        "end": "16:40",
     },
     {
-        "start": "16:35",
-        "end": "17:20",
+        "start": "16:50",
+        "end": "17:30",
     },
 ]
 
 menu_for_get_timetable = types.ReplyKeyboardMarkup(resize_keyboard=True)
-menu_for_get_timetable.row("Помощь", "Получить расписание")
+#menu_for_get_timetable.row("Помощь", "Получить расписание")
+menu_for_get_timetable.row("Получить расписание")
+menu_for_get_timetable.row("Помощь", "Настроить напоминания")
 
 menu_time_week = types.InlineKeyboardButton(
     text="На неделю",
@@ -124,6 +140,8 @@ menu_admin.row("Помощь", "Получить расписание")
 menu_admin.row("Добавить админа", "Обновить расписание")
 
 async def saveData():
+    # TODO: СДЕЛАТЬ ЧТО-ТО с data, которая уже не используется
+    return None
     dir = 'student_timetables'
     shutil.rmtree(dir)
     os.mkdir(dir)
@@ -139,6 +157,40 @@ async def saveData():
     elif system == 'Linux':
         df.to_excel(r'.\student_timetables\timetables.xlsx', sheet_name='1', index=False)
     logger.info('Обновлено расписание')
+
+def generate_schedule(key, schedule_data):
+    result = []
+    for day in range(5):
+        if key[0] in "0123456789":
+            result_day = f"\U0001F514 {key.upper()} {days[day]}: \n\n"
+        else:
+            result_day = f"\U0001F514 {days[day]}:\n\n"
+        for index, lesson in enumerate(schedule_data[day]):
+            letter = "{} - {}:    {}\n"
+            result_day += letter.format(
+                time_of_lesson[index]['start'],
+                time_of_lesson[index]['end'],
+                lesson
+            )
+        result.append(result_day)
+    return result
+def update_schedule(schedule_data):
+    global connection
+    global cursor_for_update
+    # для каждой записи в словаре, который содержит ключ class_or_last_name и список расписаний для каждого дня недели, добавляем или обновляем запись в таблице schedules
+    for k, v in schedule_data.items():
+        schedule = generate_schedule(k, v)
+        # проверяем, существует ли запись с class_or_last_name в таблице, обновляем строку или выполняем вставку в зависимости от результата
+        cursor_for_update.execute(f"SELECT COUNT(*) FROM schedules WHERE class_or_last_name=?", (k,))
+        if cursor_for_update.fetchone()[0] > 0:
+            # обновляем 5 полей расписаний в существующей строке
+            cursor_for_update.execute(f"UPDATE schedules SET Monday=?, Tuesday=?, Wednesday=?, Thursday=?, Friday=? WHERE class_or_last_name=?", (*schedule, k))
+        else:
+            # вставляем новую строку в таблицу со значениями class_or_last_name и пять полями 
+            cursor_for_update.execute(f"INSERT INTO schedules (class_or_last_name, Monday, Tuesday, Wednesday, Thursday, Friday) VALUES (?, ?, ?, ?, ?, ?)", (k, *schedule))
+
+    # сохраняем изменения
+    connection.commit()
 
 @dp.message_handler(commands="start")
 async def callback_start(message: types.Message):
@@ -171,7 +223,8 @@ async def callback_help(message: types.Message):
                            "Список команд\n"
                            "Зарегистрироваться заново - /relog\n"
                            "Открыть меню - /menu\n"
-                           "Открыть расписание - вписать 'Получить расписание'")
+                           "Открыть расписание - вписать 'Получить расписание'\n"
+                           "Отключить уведомления - /disable")
 
 
 def get_user(user_id):
@@ -182,6 +235,7 @@ def get_user(user_id):
     users = execute_read_query(connection, select_users)
 
     for user in users:
+        print(user)
         if int(user[1]) == user_id:
             return_user = {}
             for i in range(len(user)):
@@ -254,7 +308,7 @@ async def create_user(user_id, number=0, word="", is_teacher=0, teacher_last_nam
 
 
 
-
+'''
 async def get_timetable(user_id, day):
     """
         Получить расписание
@@ -288,8 +342,32 @@ async def get_timetable(user_id, day):
         logger.exception(f'Ошибка при попытке получить расписания. Пользователь: {get_user(user_id)}, Оишбка: {ex}')
         return "Ошибка... Проверьте введенный класс" \
                " (Пропишите /relog и введите нужный класс)"
+'''
+async def get_schedule_from_day(user_id, day):
+    global connection
+    global cursor_for_update
+    # создаем список дней недели
+    days_for_sqlite = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday']
 
+    # получаем соответствующее поле расписания по выбранному дню
+    day_field = days_for_sqlite[day]
+    
+    user_data = get_user(user_id)
+    if user_data['is_teacher'] == 1:
+        cursor_for_update.execute(f"SELECT {day_field} FROM schedules WHERE class_or_last_name = ?", (f"{user_data['teacher_last_name']}".lower(),))
+    else:
+        cursor_for_update.execute(f"SELECT {day_field} FROM schedules WHERE class_or_last_name = ?", (f"{user_data['number']}{user_data['word']}".lower(),))
+    
+    # выполняем запрос к базе данных, выбирая расписание для определенного класса или фамилии и для указанного дня
+    row = cursor_for_update.fetchone()
 
+    # возвращаем расписание для выбранного класса/фамилии и выбранного дня
+    if row is not None:
+        return row[0]
+    else:
+        logger.exception(f'Ошибка при попытке получить расписания. Пользователь: {get_user(user_id)}')
+        return "Ошибка... Проверьте введенный класс" \
+               " (Пропишите /relog и введите нужный класс)"
 
 
 async def start_input_person(message):
@@ -441,17 +519,17 @@ async def callback_time(call: types.CallbackQuery):
         if is_today == 1:
             day = datetime.datetime.today().weekday()
             await bot.send_message(call.from_user.id,
-                                   await get_timetable(call.from_user.id, day))
+                                   await get_schedule_from_day(call.from_user.id, day))
         elif is_today == 2:
             day = datetime.datetime.today().weekday() + 1
             if day > 4:
                 day = 0
             await bot.send_message(call.from_user.id,
-                                   await get_timetable(call.from_user.id, day))
+                                   await get_schedule_from_day(call.from_user.id, day))
         else:
             for day in range(5):
                 await bot.send_message(call.from_user.id,
-                                       await get_timetable(call.from_user.id, day))
+                                       await get_schedule_from_day(call.from_user.id, day))
 
 
 @dp.callback_query_handler(text_contains='person_')
@@ -462,10 +540,10 @@ async def callback_person(call: types.CallbackQuery):
     """
     if call.data and call.data.startswith("person_"):
         callback_get_person = call.data.split('_')[1]
-        is_today = ['student', 'teacher'].index(callback_get_person)
-        if is_today == 0:
+        is_teacher = ['student', 'teacher'].index(callback_get_person)
+        if is_teacher == 0:
             await start_input_class(call)
-        if is_today == 1:
+        if is_teacher == 1:
             await start_input_last_name(call)
 
 
@@ -536,16 +614,19 @@ async def load_timetable_file(message: types.Message, state: FSMContext):
     async with state.proxy() as update_file:
         try:
             dir = 'update_timetable'
-            shutil.rmtree(dir)
-            if document == message.document:
-                await document.download(
-                    destination_dir=dir,
-                )
+            if os.path.exists(dir):
+                shutil.rmtree(dir)
+            document = message.document
+            await document.download(
+                destination_dir=dir,
+            )
 
             if system == 'Windows':
-                data.update(get_data_students("update_timetable/documents"))
+                update_schedule(get_data_students("update_timetable/documents"))
+                #data.update(get_data_students("update_timetable/documents"))
             elif system == 'Linux':
-                data.update(get_data_students(r"update_timetable\documents"))
+                update_schedule(get_data_students(rf"update_timetable\documents"))
+                #data.update(get_data_students(rf"update_timetable\documents"))
             await saveData()
 
             await UpdateTimetable.next()
@@ -609,22 +690,187 @@ async def give_ID(message: types.Message):
     await bot.send_message(message.from_user.id, str(message.from_user.id))
 
 
+class Notification(StatesGroup):
+    """
+        Настройка напоминаний
+    """
+    time = State()
+
+@dp.message_handler(Text(equals="Настроить напоминания"))
+async def start_set_notification_time(message: types.Message):
+    """
+        Запуск настройки напоминаний
+    """
+    await start_writing_notification_time(message)
+
+async def start_writing_notification_time(message):
+    """
+        Начало ввода времени
+    """
+    await Notification.time.set()
+    await send_otmena_message(message)
+    await bot.send_message(message.from_user.id, f"Впишите время в формате HH:MM, в которое вы хотели бы получать расписание (доступно время с {interval_start} до {interval_end}, минуты будут автоматически приведены к числу кратному 5)\n\nЕсли вы хотите отключить уведомления, используйте команду /disable")
+
+def is_time_correct_string(input_string, interval_start, interval_end):
+    parts = input_string.split(':')
+    if len(parts) != 2:
+        return False
+    try:
+        hours = int(parts[0])
+        minutes = int(parts[1])
+        if hours < 0 or hours > 23:
+            return False
+        if minutes < 0 or minutes > 59:
+            return False
+        
+        start_time = datetime.strptime(interval_start, '%H:%M').time()
+        end_time = datetime.strptime(interval_end, '%H:%M').time()
+        input_time = datetime.strptime(input_string, '%H:%M').time()
+        #if '{:02d}:{:02d}'.format(hours, minutes) < interval_start or '{:02d}:{:02d}'.format(hours, minutes) > interval_end:
+        if not(start_time <= input_time <= end_time):    
+            return False
+    except ValueError:
+        return False
+    return True
+    
+def set_time(input_string):
+    parts = input_string.split(':')
+    hours = int(parts[0])
+    minutes = int(parts[1])
+    if minutes % 5 != 0:
+        minutes = (minutes // 5) * 5   # округляем минуты до ближайшего кратного 5
+    return '{:02d}:{:02d}'.format(hours, minutes)
+
+
+@dp.message_handler(state=Notification.time)
+async def set_notification_time(message: types.Message, state: FSMContext):
+    """
+        Проверка и установка времени для напоминаний
+    """
+    async with state.proxy() as new_time:
+        notification_time = message.text
+        if is_time_correct_string(notification_time, interval_start, interval_end):
+            notification_time = str(set_time(notification_time))
+            global connection
+            command = f'''
+                UPDATE
+                users
+                SET
+                notification_time = {notification_time}
+                WHERE
+                user_id = {message.from_user.id}
+            '''
+            execute_query(connection, command)
+            await bot.send_message(message.chat.id,
+                                    f'\U000023F1 Напоминания установлены на {notification_time}, для отключения введите команду /disable')
+            await bot.send_message(message.chat.id,
+                                    f'Но работать сейчас они всё равно не будут')
+            await Notification.next()
+            await state.finish()
+        else:
+            await bot.send_message(message.chat.id,
+                                   'Ошибка. Введите время в правильном формате')
+
+async def disable_notifications(user_id):
+    global connection
+    command = f'''
+        UPDATE
+        users
+        SET
+        notification_time = ""
+        WHERE
+        user_id = {user_id}
+    '''
+    execute_query(connection, command)
+    
+    
+@dp.message_handler(commands="disable")
+async def callback_menu(message: types.Message):
+    """
+        Отключить напоминания 
+    """
+    await disable_notifications(message.chat.id)
+    await bot.send_message(message.chat.id,
+                           "Напоминания отключены")
+
+
+async def send_messages():
+    now = datetime.now()
+    current_time = now.strftime("%H:%M")
+    global connection
+    global cursor_for_update
+    cursor_for_update.execute("SELECT user_id FROM users WHERE notification_time=?", (current_time,))
+    user_id_arr = [row[0] for row in cursor_for_update.fetchall()] 
+    for user_id in user_id_arr:
+        day = datetime.today().weekday()
+        await bot.send_message(user_id,
+                                await get_schedule_from_day(user_id, day))
+    
+    
+
+
+async def run_notification_schedule():
+    schedule = aioschedule.Scheduler()
+    schedule.every(1).minutes.do(send_messages)
+    start_time = datetime.now().replace(hour=interval_start_hour, minute=interval_start_minute, second=0, microsecond=0)
+    end_time = datetime.now().replace(hour=interval_end_hour, minute=interval_end_minute, second=0, microsecond=0)
+    if datetime.now() > start_time and datetime.now() < end_time:
+        while True:
+            await schedule.run_pending()
+            await asyncio.sleep(10)
+    
 async def on_startup(_):
     try:
         global connection
         connection = create_connection("db.sqlite")
-
-        global data
-        data.update(get_data_students())
-        await saveData()
-        global data_teachers
-        data_teachers = get_data_teachers()
+        
+        global cursor_for_update
+        cursor_for_update = connection.cursor()
+        
+        create_table_of_shcedule_command = '''
+            CREATE TABLE IF NOT EXISTS schedules (
+                ID INTEGER PRIMARY KEY,
+                class_or_last_name TEXT(50),
+                Monday TEXT(800),
+                Tuesday TEXT(800),
+                Wednesday TEXT(800),
+                Thursday TEXT(800),
+                Friday TEXT(800)
+            );
+        '''
+        create_table_of_users_command = """
+            CREATE TABLE IF NOT EXISTS users (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+                user_id TEXT NOT NULL,
+                number INTEGER,
+                word TEXT,
+                is_teacher INTEGER,
+                teacher_last_name TEXT,
+                is_admin INTEGER DEFAULT 0,
+                notification_time TEXT DEFAULT ""
+            );
+        """
+        execute_query(connection, create_table_of_shcedule_command)
+        execute_query(connection, create_table_of_users_command)
+        
+        
+        update_schedule(get_data_students())
+        update_schedule(get_data_teachers())
+        #await saveData()
     except Exception as ex:
         logger.exception(f'Ошибка при запуске бота: {ex}')
         sys.exit()
     else:
         logger.info('Бот запущен исправно')
-
+    '''
+    try:
+        asyncio.create_task(run_notification_schedule())
+    except Exception as ex:
+        logger.exception(f'Ошибка при запуске напоминаний: {ex}')
+        sys.exit()
+    else:
+        logger.info('Напоминания запущены исправно')
+    '''
 
 
 if __name__ == "__main__":
